Ajuste propriedade double para BigDecimal

+++ Mateus_Pfeffer/recharge-api/src/main/java/com/fag/infra/celcoin/dto/CelcoinRechargeValueDTO.java
@@ -1,4 +1,6 @@
 package com.fag.infra.celcoin.dto;
 
-public record CelcoinRechargeValueDTO(Double value) {
+import java.math.BigDecimal;
+
+public record CelcoinRechargeValueDTO(BigDecimal value) {
 }
